Opt out of edge-to-edge enforcement
This fixes screwed up UI in the short-term. The long term fix is to implement it correctly, but we have until Android 16 to do that. Since there are other issues with Android 16 (keyboard doesn't appear at all), we take the easy way out now to buy time while we assess whether the keyboard is salvageable in its current state under Android 16 at all, or whether more significant changes are required.