Upgrade to Android 15 (SDK 35) (#47)
* WIP Android 15 (SDK 35) support

* Update divvunspell and pahkat_client libs to use 16KB page size

This is required of app targeting Android 15. The libs were rebuilt by:
1. Installing NDK 28 (29 works too) via Android studio
2. Selecting it in Android Studio > File > Project Structure
3. Installing `cargo-ndk` with `cargo install cargo-ndk`
4. Checking out https://github.com/divvun/divvunspell and then inside that dir, running `cargo ndk -t armeabi-v7a -t arm64-v8a -o ./lib build -vv --lib --release --features internal_ffi`
5. Checking out https://github.com/divvun/pahkat and going to the `pahkat-client-core` dir and running `cargo ndk -t armeabi-v7a -t arm64-v8a -o ./lib build -vv --features ffi,prefix --release`
6. Moving the output of the above commands to the `jniLibs` dir in the Android project

* Remove AI garbage

* Update to gradle 8.12.0 and run upgrade assistant

* Handle edge-to-edge for User Dictionary view

* Opt out of edge-to-edge enforcement

This fixes screwed up UI in the short-term. The long term fix is to implement it correctly, but we have until Android 16 to do that. Since there are other issues with Android 16 (keyboard doesn't appear at all), we take the easy way out now to buy time while we assess whether the keyboard is salvageable in its current state under Android 16 at all, or whether more significant changes are required.

* Revert "Update divvunspell and pahkat_client libs to use 16KB page size"

This reverts commit 63b525ada129f5e709020ae692fa89ea362f516d.

* Add instructions for building DivvunSpell and Pahkat